+++ db4e/Modules/Db4E.py
@@ -2,28 +2,34 @@
 db4e/Modules/Db4E.py
 
     Database 4 Everything
-    Author: Nadim-Daniel Ghaznavi 
+    Author: Nadim-Daniel Ghaznavi
     Copyright: (c) 2024-2025 Nadim-Daniel Ghaznavi
     GitHub: https://github.com/NadimGhaznavi/db4e
     License: GPL 3.0
 
 A class representing the deployment of Db4E
 """
+
 import os, grp, getpass
 
 from db4e.Modules.LocalSoftwareSystem import LocalSoftwareSystem
 from db4e.Modules.Components import (
-    DonationWallet, Db4eGroup, InstallDir, Db4eUser, UserWallet, VendorDir,
-    PrimaryServer, Instance)
+    DonationWallet,
+    Db4eGroup,
+    InstallDir,
+    Db4eUser,
+    UserWallet,
+    VendorDir,
+    PrimaryServer,
+    Instance,
+)
 from db4e.Constants.DField import DField
 from db4e.Constants.DElem import DElem
 from db4e.Constants.DLabel import DLabel
 from db4e.Constants.DDef import DDef
 
 
-
 class Db4E(LocalSoftwareSystem):
-
 
     def __init__(self, rec=None):
         super().__init__()
@@ -38,7 +44,7 @@ class Db4E(LocalSoftwareSystem):
         self.add_component(DField.USER, Db4eUser())
         self.add_component(DField.USER_WALLET, UserWallet())
         self.add_component(DField.VENDOR_DIR, VendorDir())
-        
+
         self.donation_wallet = self.components[DField.DONATION_WALLET]
         self.group = self.components[DField.GROUP]
         self.install_dir = self.components[DField.INSTALL_DIR]
@@ -59,12 +65,23 @@ class Db4E(LocalSoftwareSystem):
         if rec:
             self.from_rec(rec)
 
+    def __dict__(self):
+        db4e_dict = {
+            DField.DONATION_WALLET: self.donation_wallet(),
+            DField.DB4E_GROUP: self.group(),
+            DField.DB4E_USER: self.user(),
+            DField.INSTALL_DIR: self.install_dir(),
+            DField.INSTANCE: self.instance(),
+            DField.PRIMARY_SERVER: self.primary_server(),
+            DField.USER_WALLET: self.user_wallet(),
+            DField.VENDOR_DIR: self.vendor_dir(),
+        }
+        return db4e_dict
 
     def instance_map(self, map=None):
         if map:
             self._instance_map = map
         return self._instance_map
-
 
     def set_effective_identity(self):
         """Set the Db4E user and group based on who is running this app"""
@@ -77,6 +94,5 @@ class Db4E(LocalSoftwareSystem):
         self.user(user)
         self.group(group)
 
-
     def set_install_dir(self):
-        self.install_dir(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+        self.install_dir(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))