+++ db4e/Modules/MoneroDRemote.py
@@ -2,7 +2,7 @@
 db4e/Modules/MoneroDRemote.py
 
     Database 4 Everything
-    Author: Nadim-Daniel Ghaznavi 
+    Author: Nadim-Daniel Ghaznavi
     Copyright: (c) 2024-2025 Nadim-Daniel Ghaznavi
     GitHub: https://github.com/NadimGhaznavi/db4e
     License: GPL 3.0
@@ -12,15 +12,19 @@ Everything remote Monero Daemon
 
 from db4e.Modules.SoftwareSystem import SoftwareSystem
 from db4e.Modules.Components import (
-    Instance, Remote, RpcBindPort, IpAddr, ZmqPubPort, PrimaryServer)
+    Instance,
+    Remote,
+    RpcBindPort,
+    IpAddr,
+    ZmqPubPort,
+    PrimaryServer,
+)
 from db4e.Constants.DLabel import DLabel
 from db4e.Constants.DField import DField
 from db4e.Constants.DElem import DElem
 
 
-
 class MoneroDRemote(SoftwareSystem):
-
 
     def __init__(self, rec=None):
         super().__init__()
@@ -42,3 +46,11 @@ class MoneroDRemote(SoftwareSystem):
         if rec:
             self.from_rec(rec)
 
+    def __dict__(self):
+        monerod_remote_dict = {
+            DField.INSTANCE: self.instance.value(),
+            DField.RPC_BIND_PORT: self.rpc_bind_port.value(),
+            DField.IP_ADDR: self.ip_addr.value(),
+            DField.ZMQ_PUB_PORT: self.zmq_pub_port.value(),
+        }
+        return monerod_remote_dict