+++ db4e/Modules/P2PoolRemote.py
@@ -2,7 +2,7 @@
 db4e/Modules/P2PoolRemote.py
 
     Database 4 Everything
-    Author: Nadim-Daniel Ghaznavi 
+    Author: Nadim-Daniel Ghaznavi
     Copyright: (c) 2024-2025 Nadim-Daniel Ghaznavi
     GitHub: https://github.com/NadimGhaznavi/db4e
     License: GPL 3.0
@@ -11,15 +11,13 @@ Everything P2Pool Remote
 """
 
 from db4e.Modules.SoftwareSystem import SoftwareSystem
-from db4e.Modules.Components import (Instance, Remote, IpAddr, StratumPort)
+from db4e.Modules.Components import Instance, Remote, IpAddr, StratumPort
 from db4e.Constants.DElem import DElem
 from db4e.Constants.DField import DField
 from db4e.Constants.DLabel import DLabel
 
 
-
 class P2PoolRemote(SoftwareSystem):
-
 
     def __init__(self, rec=None):
         super().__init__()
@@ -39,4 +37,10 @@ class P2PoolRemote(SoftwareSystem):
         if rec:
             self.from_rec(rec)
 
-
+    def __dict__(self):
+        p2pool_remote_dict = {
+            DField.INSTANCE: self.instance(),
+            DField.IP_ADDR: self.ip_addr(),
+            DField.STRATUM_PORT: self.stratum_port(),
+        }
+        return p2pool_remote_dict