+++ db4e/server.py
@@ -2,7 +2,7 @@
 db4e/server.py
 
     Database 4 Everything
-    Author: Nadim-Daniel Ghaznavi 
+    Author: Nadim-Daniel Ghaznavi
     Copyright: (c) 2024-2025 Nadim-Daniel Ghaznavi
     GitHub: https://github.com/NadimGhaznavi/db4e
     License: GPL 3.0
@@ -10,6 +10,7 @@ db4e/server.py
 """
 
 import os, sys
+
 # Turn off buffering
 sys.stdout.reconfigure(line_buffering=True)
 from datetime import datetime
@@ -47,6 +48,7 @@ from db4e.Modules.OpsDb import OpsDb, OpsETL
 from db4e.Modules.P2Pool import P2Pool
 from db4e.Modules.P2PoolRemote import P2PoolRemote
 from db4e.Modules.P2PoolWatcher import P2PoolWatcher
+from db4e.Modules.SQLMgr import SQLMgr
 from db4e.Modules.XMRig import XMRig
 
 from db4e.Constants.DDebug import DDebug
@@ -63,10 +65,10 @@ from db4e.Constants.DLabel import DLabel
 from db4e.Constants.DMongo import DMongo
 
 
-
 DDebug.FUNCTION = False
 
 POLL_INTERVAL = 5
+
 
 class Db4eServer:
     """
@@ -75,12 +77,13 @@ class Db4eServer:
     https://db4e.osoyalce.com/images/Server-Relationships.png
     """
 
-
-
     def __init__(self):
 
         # Mongo DB Manager
         self.db = DbMgr()
+
+        # SQLite DB Manager
+        self.sqldb = SQLMgr(db_type=DField.SERVER)
 
         #  systemd wrapper
         self.systemd = Db4ESystemD(db=self.db)
@@ -93,18 +96,19 @@ class Db4eServer:
 
         # Job Queue
         self.job_queue = JobQueue(db=self.db)
-        
+
         # Setup logging
         vendor_dir = self.depl_mgr.get_dir(DDir.VENDOR)
         logs_dir = DDef.LOG_DIR
         log_file = DDef.DB4E_LOG_FILE
         fq_log_file = os.path.join(vendor_dir, DElem.DB4E, logs_dir, log_file)
-        self.log_file = fq_log_file 
+        self.log_file = fq_log_file
         self.log = Db4ELogger(db4e_module=DModule.DB4E_SERVER, log_file=fq_log_file)
 
-        # Mining DB object 
+        # Mining DB object
         self.mining_db = MiningDb(
-            db=self.db, log_file=fq_log_file, ops_etl=OpsETL(ops_db=OpsDb(db=self.db)))
+            db=self.db, log_file=fq_log_file, ops_etl=OpsETL(ops_db=OpsDb(db=self.db))
+        )
 
         # Database Cache
         self.db_cache = DbCache(db=self.db, mining_db=self.mining_db)
@@ -130,7 +134,6 @@ class Db4eServer:
         # Make sure the permissions on the logrotate files are correct
         self.chown_logrotate_files()
 
-
     def add_deployment(self, job):
         if DDebug.FUNCTION:
             self.log.debug(f"Db4eServer:add_deployment(): {job}")
@@ -141,7 +144,6 @@ class Db4eServer:
         job.elem_type(elem.elem_type())
         job.instance(elem.instance())
         self.job_queue.complete_job(job)
-
 
     def check_deployments(self):
         depls = self.depl_mgr.get_deployments()
@@ -154,12 +156,12 @@ class Db4eServer:
 
             # Nothing to do for these classes
             if elem_type == P2PoolRemote:
-                continue 
+                continue
 
             # Look for primary Monero deployments
             if elem_type == Db4E:
                 primary_server = elem.primary_server()
-                #print(f"Db4eServer:check_deployments(): primary_server: {primary_server}")
+                # print(f"Db4eServer:check_deployments(): primary_server: {primary_server}")
                 if primary_server == DField.DISABLE:
                     self.unset_int_p2pool_primary()
                 else:
@@ -171,18 +173,20 @@ class Db4eServer:
             if elem_type in [MoneroD, P2Pool, InternalP2Pool, XMRig] and elem.enabled():
                 self.ensure_running(elem)
                 if elem_type in [P2Pool, InternalP2Pool]:
-                    # Make sure there's a log watcher running                    
-                    self.spawn_log_watcher(elem) 
+                    # Make sure there's a log watcher running
+                    self.spawn_log_watcher(elem)
 
             # Makre sure anything that's disabled is stopped
-            #self.log.debug(f"Db4eServer:check_deployments(): enabled: {elem}: {elem.enabled()}")
-            if elem_type in [MoneroD, P2Pool, InternalP2Pool, XMRig] and not elem.enabled():
+            # self.log.debug(f"Db4eServer:check_deployments(): enabled: {elem}: {elem.enabled()}")
+            if (
+                elem_type in [MoneroD, P2Pool, InternalP2Pool, XMRig]
+                and not elem.enabled()
+            ):
                 self.ensure_stopped(elem)
 
             # There are no primary Monery deployments
             if not found_primary:
                 self.unset_int_p2pool_primary()
-
 
     def check_jobs(self):
         if DDebug.FUNCTION:
@@ -195,9 +199,9 @@ class Db4eServer:
                 jobs.append(job)
             else:
                 found_job = False
-        
+
         for job in jobs:
-            #self.log.debug(f"Db4eServer:check_jobs(): job.elem(): {job.elem()}")
+            # self.log.debug(f"Db4eServer:check_jobs(): job.elem(): {job.elem()}")
             op = job.op()
             if op == DJob.NEW:
                 self.add_deployment(job=job)
@@ -215,7 +219,6 @@ class Db4eServer:
                 self.set_primary(job=job)
             self.job_queue.complete_job(job)
 
-
     def chown_logrotate_files(self):
         if DDebug.FUNCTION:
             self.log.debug("Db4eServer:chown_logrotate_files():")
@@ -226,15 +229,11 @@ class Db4eServer:
             fq_file = os.path.join(logrotate_dir, aFile)
             try:
                 cmd = [DFile.SUDO, DFile.CHOWN, DDef.ROOT, fq_file]
-                proc = subprocess.run(
-                    cmd,
-                    stderr=subprocess.PIPE,
-                    input="")
-                stderr = proc.stderr.decode('utf-8')
+                proc = subprocess.run(cmd, stderr=subprocess.PIPE, input="")
+                stderr = proc.stderr.decode("utf-8")
                 self.log.info(f"Set permissions on logrotate file: {fq_file}")
             except Exception as e:
                 self.log.critical(f"chown_logrotate_files() failed: {e} {stderr}")
-
 
     def delete(self, job: Job):
         if DDebug.FUNCTION:
@@ -270,7 +269,6 @@ class Db4eServer:
             self.ensure_stopped(elem)
             self.depl_mgr.delete_deployment(elem)
             self.job_queue.complete_job(job=job)
-            
 
     def disable(self, job: Job):
         if DDebug.FUNCTION:
@@ -279,23 +277,30 @@ class Db4eServer:
         instance = job.instance()
         elem = self.depl_mgr.get_deployment(elem_type, instance)
 
-        #print(f"Db4eServer:disable(): {elem}: current: {elem.enabled()}")
+        # print(f"Db4eServer:disable(): {elem}: current: {elem.enabled()}")
         if not elem.enabled():
             return
         job.msg(f"Disabled deployment")
         elem.enabled(False)
         self.depl_mgr.update_deployment(elem)
         self.job_queue.complete_job(job)
-        if type(elem) == P2Pool or type(elem) == MoneroD or \
-            type(elem) == P2PoolRemote or type(elem) == MoneroDRemote:
+        if (
+            type(elem) == P2Pool
+            or type(elem) == MoneroD
+            or type(elem) == P2PoolRemote
+            or type(elem) == MoneroDRemote
+        ):
             self.disable_downstream(elem)
         self.log.info(f"Disable: {elem}")
         # Create an Ops record when remote elements are disabled
         if elem_type == DElem.MONEROD_REMOTE:
-            self.ops_db.add_stop_event(elem_type=DLabel.MONEROD_REMOTE_SHORT, instance=instance)
-        elif  elem_type == DElem.P2POOL_REMOTE:
-            self.ops_db.add_stop_event(elem_type=DLabel.P2POOL_REMOTE_SHORT, instance=instance)
-
+            self.ops_db.add_stop_event(
+                elem_type=DLabel.MONEROD_REMOTE_SHORT, instance=instance
+            )
+        elif elem_type == DElem.P2POOL_REMOTE:
+            self.ops_db.add_stop_event(
+                elem_type=DLabel.P2POOL_REMOTE_SHORT, instance=instance
+            )
 
     def disable_downstream(self, elem):
         if DDebug.FUNCTION:
@@ -310,7 +315,11 @@ class Db4eServer:
                     self.ensure_stopped(p2pool)
                     self.depl_mgr.update_deployment(p2pool)
                     p2pool.parent(DField.DISABLE)
-                    job = Job(op=DJob.DISABLE, elem_type=elem.elem_type(), instance=elem.instance())
+                    job = Job(
+                        op=DJob.DISABLE,
+                        elem_type=elem.elem_type(),
+                        instance=elem.instance(),
+                    )
                     job.msg(f"Disabled downstream instance: {elem.instance()}")
                     self.disable_downstream(p2pool)
 
@@ -323,7 +332,11 @@ class Db4eServer:
                     self.ensure_stopped(int_p2pool)
                     self.depl_mgr.update_deployment(int_p2pool)
                     int_p2pool.parent(DField.DISABLE)
-                    job = Job(op=DJob.DISABLE, elem_type=elem.elem_type(), instance=elem.instance())
+                    job = Job(
+                        op=DJob.DISABLE,
+                        elem_type=elem.elem_type(),
+                        instance=elem.instance(),
+                    )
                     job.msg(f"Disabled downstream instance: {elem.instance()}")
 
         elif type(elem) == P2Pool or type(elem) == P2PoolRemote:
@@ -335,10 +348,13 @@ class Db4eServer:
                     xmrig.enabled(False)
                     xmrig.parent(DField.DISABLE)
                     self.depl_mgr.update_deployment(xmrig)
-                    job = Job(op=DJob.DISABLE, elem_type=elem.elem_type(), instance=elem.instance())
+                    job = Job(
+                        op=DJob.DISABLE,
+                        elem_type=elem.elem_type(),
+                        instance=elem.instance(),
+                    )
                     job.msg(f"Disabled downstream instance: {elem.instance()}")
-                    self.job_queue.complete_job(job)    
-
+                    self.job_queue.complete_job(job)
 
     def enable(self, job: Job):
         if DDebug.FUNCTION:
@@ -346,22 +362,24 @@ class Db4eServer:
         elem_type = job.elem_type()
         instance = job.instance()
         elem = self.depl_mgr.get_deployment(elem_type, instance)
-        #print(f"Db4eServer:enable(): {elem}: current: {elem.enabled()}")
+        # print(f"Db4eServer:enable(): {elem}: current: {elem.enabled()}")
         if elem.enabled():
             return
         self.log.info(f"Enable: {elem}")
-        
+
         job.msg(f"Enabled deployment")
         elem.enabled(True)
         self.depl_mgr.update_deployment(elem)
         self.job_queue.complete_job(job)
         # Create an Ops record when remote elements are enabled
         if elem_type == DElem.MONEROD_REMOTE:
-            self.ops_db.add_start_event(elem_type=DLabel.MONEROD_REMOTE_SHORT, instance=instance)
-        elif  elem_type == DElem.P2POOL_REMOTE:
-            self.ops_db.add_start_event(elem_type=DLabel.P2POOL_REMOTE_SHORT, instance=instance)
-
-
+            self.ops_db.add_start_event(
+                elem_type=DLabel.MONEROD_REMOTE_SHORT, instance=instance
+            )
+        elif elem_type == DElem.P2POOL_REMOTE:
+            self.ops_db.add_start_event(
+                elem_type=DLabel.P2POOL_REMOTE_SHORT, instance=instance
+            )
 
     def ensure_running(self, elem):
         if DDebug.FUNCTION:
@@ -370,13 +388,13 @@ class Db4eServer:
         sd = self.systemd
         if type(elem) == MoneroD:
             instance = elem.instance()
-            sd.service_name('monerod@' + instance)
+            sd.service_name("monerod@" + instance)
         elif type(elem) == P2Pool or type(elem) == InternalP2Pool:
             instance = elem.instance()
-            sd.service_name('p2pool@' + instance)
+            sd.service_name("p2pool@" + instance)
         elif type(elem) == XMRig:
             instance = elem.instance()
-            sd.service_name('xmrig@' + instance)
+            sd.service_name("xmrig@" + instance)
         else:
             raise ValueError(f"Unknown deployment type: {elem}")
 
@@ -393,13 +411,12 @@ class Db4eServer:
         self.starting.add(instance)
         rc = sd.start()
         if rc == 0:
-            self.log.info(f'Started: {elem}')
+            self.log.info(f"Started: {elem}")
             time.sleep(30)
         else:
-            self.log.critical(f'ERROR: Failed to start {elem}, return code was {rc}')
+            self.log.critical(f"ERROR: Failed to start {elem}, return code was {rc}")
             self.stopping.discard(instance)
-            self.starting.discard(instance)            
-
+            self.starting.discard(instance)
 
     def ensure_stopped(self, elem):
         if DDebug.FUNCTION:
@@ -407,13 +424,13 @@ class Db4eServer:
         sd = self.systemd
         if type(elem) == MoneroD:
             instance = elem.instance()
-            sd.service_name('monerod@' + instance)
+            sd.service_name("monerod@" + instance)
         elif isinstance(elem, P2Pool):
             instance = elem.instance()
-            sd.service_name('p2pool@' + instance)
+            sd.service_name("p2pool@" + instance)
         elif type(elem) == XMRig:
             instance = elem.instance()
-            sd.service_name('xmrig@' + instance)
+            sd.service_name("xmrig@" + instance)
         else:
             raise ValueError(f"Unknown deployment type: {elem}")
 
@@ -426,12 +443,12 @@ class Db4eServer:
         if instance in self.stopping:
             # It's already in the process of stopping, do nothing
             return
-        
+
         # Active and not already stopping -> issue stop
         self.stopping.add(instance)
         rc = sd.stop()
         if rc == 0:
-            self.log.info(f'Stopped: {elem}')
+            self.log.info(f"Stopped: {elem}")
             time.sleep(30)
             if isinstance(elem, P2Pool):
                 control = self.log_watchers.pop(instance, None)
@@ -441,8 +458,7 @@ class Db4eServer:
                     stop_event.set()
                     thread.join()
         else:
-            self.log.critical(f'ERROR: Failed to stop {elem}, return code was {rc}')
-
+            self.log.critical(f"ERROR: Failed to stop {elem}, return code was {rc}")
 
     def mongodb_backup(self):
         if DDebug.FUNCTION:
@@ -452,12 +468,11 @@ class Db4eServer:
         jobs_col = DDef.JOBS_COLLECTION
         mining_col = DDef.MINING_COLLECTION
         ops_col = DDef.OPS_COLLECTION
-        all_cols = [ depl_col, jobs_col, mining_col, ops_col ]
+        all_cols = [depl_col, jobs_col, mining_col, ops_col]
 
         now = datetime.now()
         timestamp = now.strftime("%Y-%m-%d")
-        todays_backup = \
-            timestamp + "_" + DElem.DB4E + '_' + depl_col + DDef.GZIP_SUFFIX
+        todays_backup = timestamp + "_" + DElem.DB4E + "_" + depl_col + DDef.GZIP_SUFFIX
         vendor_dir = self.depl_mgr.get_dir(DDir.VENDOR)
         backup_dir = os.path.join(vendor_dir, DDef.BACKUP_DIR)
         fq_todays_backup = os.path.join(backup_dir, todays_backup)
@@ -465,7 +480,7 @@ class Db4eServer:
         # Check if today's backup already exists
         if os.path.exists(fq_todays_backup):
             return
-        
+
         # Create the backup directory if it doesn't exist
         if not os.path.isdir(backup_dir):
             os.makedirs(backup_dir)
@@ -474,26 +489,26 @@ class Db4eServer:
         try:
             for aCol in all_cols:
                 dumpfile = os.path.join(
-                    backup_dir, timestamp + "_" + DElem.DB4E + '_' + aCol)
+                    backup_dir, timestamp + "_" + DElem.DB4E + "_" + aCol
+                )
                 dumpfile_gz = dumpfile + DDef.GZIP_SUFFIX
 
                 ## Backup Mongo; run mongodump
-                cmd = [ 
-                    DFile.MONGODUMP, 
+                cmd = [
+                    DFile.MONGODUMP,
                     f"--archive={dumpfile}",
                     f"--db={DElem.DB4E}",
-                    f"--collection={aCol}"
-                    ]
+                    f"--collection={aCol}",
+                ]
                 proc = subprocess.run(
-                    cmd,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE)
+                    cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+                )
                 stdout = proc.stdout.decode()
                 stderr = proc.stderr.decode()
 
                 ## Compress the dump files
-                with open(dumpfile, 'rb') as f_in:
-                    with gzip.open(dumpfile_gz, 'wb') as f_out:
+                with open(dumpfile, "rb") as f_in:
+                    with gzip.open(dumpfile_gz, "wb") as f_out:
                         # Write the contents of the original file to the gzipped file
                         shutil.copyfileobj(f_in, f_out)
                 os.remove(dumpfile)
@@ -508,7 +523,7 @@ class Db4eServer:
             # We only keep MAX_LOG_FILES backups for each backup file
             if len(dump_files) <= len(all_cols) * DDef.MAX_LOG_FILES:
                 return
-            
+
             # Collect creation times
             dumps_with_times = []
             for aDump in dump_files:
@@ -525,11 +540,9 @@ class Db4eServer:
                 os.remove(dump_path)
                 self.log.info(f"Deleted old backup file: {dump_path}")
 
-
         except Exception as e:
             self.log.critical(f"mongodb_backup() failed: {e} {stdout} {stderr}")
             return
-
 
     def restart(self, job):
         if DDebug.FUNCTION:
@@ -540,15 +553,14 @@ class Db4eServer:
         instance = job.instance()
         sd = self.systemd
         if elem_type == DElem.MONEROD:
-            sd.service_name('monerod@' + instance)
+            sd.service_name("monerod@" + instance)
         elif elem_type == DElem.P2POOL:
-            sd.service_name('p2pool@' + instance)
+            sd.service_name("p2pool@" + instance)
         else:
             raise ValueError(f"Unknown deployment type: {elem_type}")
         sd.restart()
         job.msg(f"Restarted instance")
         self.job_queue.complete_job(job)
-
 
     def rotate_logs(self):
         # Run logrotate every two hours
@@ -558,78 +570,79 @@ class Db4eServer:
             self.last_logrotate = cur_hour
             try:
                 logrotate_dir = self.depl_mgr.get_dir(DDir.LOGROTATE)
-                cmd = [DFile.SUDO , DFile.LOGROTATE, "-v", logrotate_dir]
+                cmd = [DFile.SUDO, DFile.LOGROTATE, "-v", logrotate_dir]
                 proc = subprocess.run(
-                    cmd,
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.PIPE,
-                    input='')
+                    cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, input=""
+                )
                 stdout = proc.stdout.decode()
                 stderr = proc.stderr.decode()
-                #self.log.debug(f"rotate_logs(): {stdout}{stderr}")
-                
+                # self.log.debug(f"rotate_logs(): {stdout}{stderr}")
+
                 output_lines = stderr.split("\n")
                 depls = {}
                 cur_elem_type = None
                 cur_instance = None
                 for line in output_lines:
-                    #self.log.critical(line)
+                    # self.log.critical(line)
                     # Parse the elem_type and instance out of the logrotate output
                     pattern = r".*reading config file\s(?P<config>.*.conf)"
                     match = re.search(pattern, line)
                     if match:
-                        config_file = match.group('config')
-                        #self.log.critical(f"Found config: {config_file}")
+                        config_file = match.group("config")
+                        # self.log.critical(f"Found config: {config_file}")
                         if config_file == DElem.DB4E + DDef.CONF_SUFFIX:
                             continue
                         else:
                             pattern = r"(?P<elem_type>.*)-(?P<instance>.*).conf"
                             match = re.search(pattern, config_file)
                             if match:
-                                elem_type = match.group('elem_type')
-                                instance = match.group('instance')
-                                #self.log.critical(f"Found {elem_type}/{instance}")
+                                elem_type = match.group("elem_type")
+                                instance = match.group("instance")
+                                # self.log.critical(f"Found {elem_type}/{instance}")
                                 depls[(elem_type, instance)] = False
-                    
+
                     # Watch to see if a log was rotated
                     pattern = rf"considering log {re.escape(vendor_dir)}/(?P<elem_type>[^/]+)(?:/(?P<instance>[^/]+))?/logs/(?P<logname>[^/]+)\.log"
                     match = re.search(pattern, line)
                     if match:
-                        #self.log.critical(f"line: {line}")
-                        cur_elem_type = match.group('elem_type')
-                        cur_instance = match.group('instance') or match.group('logname')
-                        #self.log.critical(f"{cur_elem_type}/{cur_instance}")
+                        # self.log.critical(f"line: {line}")
+                        cur_elem_type = match.group("elem_type")
+                        cur_instance = match.group("instance") or match.group("logname")
+                        # self.log.critical(f"{cur_elem_type}/{cur_instance}")
                     pattern = r"\s+log needs rotating.*"
                     match = re.search(pattern, line)
                     if match:
                         if cur_elem_type != DElem.DB4E:
                             depls[(cur_elem_type, cur_instance)] = True
-                            self.log.info(f"{cur_elem_type}/{cur_instance} rotating log file")
+                            self.log.info(
+                                f"{cur_elem_type}/{cur_instance} rotating log file"
+                            )
 
                 for elem_type, instance in depls:
                     if depls[(elem_type, instance)]:
                         # Create a restart job
-                        job = Job(op=DJob.RESTART, elem_type=elem_type, instance=instance)
+                        job = Job(
+                            op=DJob.RESTART, elem_type=elem_type, instance=instance
+                        )
                         self.job_queue.post_job(job)
-
 
             except Exception as e:
                 self.log.error(f"rotate_logs(): {e} {stderr}")
                 return
-            
 
     def shutdown(self, signum, frame):
         if DDebug.FUNCTION:
             self.log.debug(f"Db4eServer:shutdown(): {signum}")
-        self.log.info(f'Shutdown requested (signal {signum})')
+        self.log.info(f"Shutdown requested (signal {signum})")
         self.running.clear()
         for instance in self.log_watchers.keys():
-            self.ops_db.add_stop_event(elem_type=DLabel.P2POOL_WATCHER, instance=instance)
+            self.ops_db.add_stop_event(
+                elem_type=DLabel.P2POOL_WATCHER, instance=instance
+            )
 
         # Create a stop event in the ops collection
         self.ops_db.add_stop_event(elem_type=DLabel.DB4E, instance=DElem.DB4E)
         sys.exit(0)
-
 
     def set_int_p2pool_primary(self, monerod_id):
         if DDebug.FUNCTION:
@@ -644,12 +657,16 @@ class Db4eServer:
                 p2pool.gen_config(tmpl_file=tmpl_file, vendor_dir=vendor_dir)
                 p2pool.log_file(
                     os.path.join(
-                        vendor_dir, self.depl_mgr.get_dir(DElem.P2POOL), p2pool.instance(), 
-                        DDir.LOG, DFile.P2POOL_LOG))
+                        vendor_dir,
+                        self.depl_mgr.get_dir(DElem.P2POOL),
+                        p2pool.instance(),
+                        DDir.LOG,
+                        DFile.P2POOL_LOG,
+                    )
+                )
                 p2pool.enabled(True)
                 self.depl_mgr.update_deployment(p2pool)
                 self.ensure_running(p2pool)
-
 
     def spawn_log_watcher(self, p2pool):
         if DDebug.FUNCTION:
@@ -686,7 +703,8 @@ class Db4eServer:
             )
         else:
             raise ValueError(
-                f"spawn_log_watcher(): Unknown deployment type: {type(p2pool)}")
+                f"spawn_log_watcher(): Unknown deployment type: {type(p2pool)}"
+            )
 
         def _runner():
             try:
@@ -698,18 +716,20 @@ class Db4eServer:
                 watcher.stop_sub_thread()
                 self.log_watchers.pop(instance, None)
                 self.log.debug(f"Watcher thread exiting: {instance}")
-                if instance not in [ DLabel.MAIN_CHAIN, DLabel.MINI_CHAIN, 
-                                    DLabel.NANO_CHAIN]:
+                if instance not in [
+                    DLabel.MAIN_CHAIN,
+                    DLabel.MINI_CHAIN,
+                    DLabel.NANO_CHAIN,
+                ]:
                     self.ops_db.add_start_event(
-                        elem_type=DLabel.P2POOL_WATCHER, instance=instance)
-
+                        elem_type=DLabel.P2POOL_WATCHER, instance=instance
+                    )
 
         t = threading.Thread(target=_runner, name=f"LogWatcher-{instance}", daemon=True)
         self.log_watchers[instance] = (t, stop_event, watcher)
         t.start()
         self.log.info(f"Started P2Pool watcher: {instance}")
         self.ops_db.add_start_event(elem_type=DLabel.P2POOL_WATCHER, instance=instance)
-
 
     def start(self):
         if DDebug.FUNCTION:
@@ -727,7 +747,6 @@ class Db4eServer:
             self.mongodb_backup()
             time.sleep(POLL_INTERVAL)
 
-
     def update(self, job):
         if DDebug.FUNCTION:
             self.log.debug(f"Db4eServer:update(): {job}")
@@ -744,10 +763,11 @@ class Db4eServer:
 
         # Restart Monerod and P2Pool deployments if their config has been updated
         if type(elem) == MoneroD or type(elem) == P2Pool:
-            # Create a restart job 
-            job = Job(op=DJob.RESTART, elem_type=elem.elem_type(), instance=elem.instance())
+            # Create a restart job
+            job = Job(
+                op=DJob.RESTART, elem_type=elem.elem_type(), instance=elem.instance()
+            )
             self.job_queue.post_job(job)
-
 
     def unset_int_p2pool_primary(self):
         if DDebug.FUNCTION:
@@ -758,7 +778,6 @@ class Db4eServer:
             self.depl_mgr.update_deployment(p2pool)
 
 
-
 def main():
     # Set environment variables for better color support
     os.environ[DField.TERM_ENVIRON] = DDef.TERM
@@ -766,5 +785,7 @@ def main():
 
     server = Db4eServer()
     server.start()
+
+
 if __name__ == "__main__":
     main()